add --unlit option
Removing literate comments from a Haskell source file is certainly a
pre-processing task, but rarely implemented separately from a compiler.
But the code is so simple, it ought to be put in a standard place so
anyone can re-use it, whether they use cpp or not.  This commit exposes
the 'unlit' function as the library module Language.Preprocessor.Unlit,
as well as adding it as an option to cpphs.