Major API change: declare main functions with IO type.
To reflect their true nature, all of runCpphs, cppIfdef, and macroPass,
now return their results in the IO monad.  Previous uses of
unsafePerformIO have been removed.  (Instead, unsafeInterleaveIO is used
in an attempt to preserve lazy space behaviour, but this usage is
as safe as anything else in the IO monad.)